Move Config object to the config module and rename to AnalyticsConfig
