+++ features/analytics/impl/src/main/kotlin/io/element/android/features/analytics/impl/preferences/DefaultAnalyticsPreferencesPresenter.kt
@@ -20,8 +20,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.rememberCoroutineScope
 import com.squareup.anvil.annotations.ContributesBinding
+import io.element.android.appconfig.AnalyticsConfig
 import io.element.android.features.analytics.api.AnalyticsOptInEvents
-import io.element.android.features.analytics.api.Config
 import io.element.android.features.analytics.api.preferences.AnalyticsPreferencesPresenter
 import io.element.android.features.analytics.api.preferences.AnalyticsPreferencesState
 import io.element.android.libraries.core.meta.BuildMeta
@@ -52,7 +52,7 @@ class DefaultAnalyticsPreferencesPresenter @Inject constructor(
         return AnalyticsPreferencesState(
             applicationName = buildMeta.applicationName,
             isEnabled = isEnabled.value,
-            policyUrl = Config.POLICY_LINK,
+            policyUrl = AnalyticsConfig.POLICY_LINK,
             eventSink = ::handleEvents
         )
     }
